requirements.txt
# MCP SDK (includes FastMCP)
mcp>=1.0.0

# HTTP client (Server A → Server B)
httpx>=0.27.0

# Web framework (Server B HTTP endpoints)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# OpenAI SDK
openai>=1.10.0

# Data validation
pydantic>=2.5.0

# Environment variable loading
python-dotenv>=1.0.0
